@@ -1,6 +1,6 @@
 {
 	"name": "skynovel",
-	"version": "3.12.5",
+	"version": "3.12.6",
 	"publisher": "famibee",
 	"displayName": "SKYNovel",
 	"description": "It supports development of SKYNovel for VSCode, VSCodeでSKYNovelの開発をサポートする拡張機能",
@@ -9,7 +9,7 @@
 	"homepage": "http://famibee.blog38.fc2.com/",
 	"license": "MIT",
 	"engines": {
-		"vscode": "^1.55.0"
+		"vscode": "^1.56.0"
 	},
 	"categories": [
 		"Programming Languages",
@@ -664,10 +664,10 @@
 		"await-notify": "^1.0.1",
 		"crc-32": "^1.2.0",
 		"crypto-js": "^4.0.0",
-		"fs-extra": "^9.1.0",
+		"fs-extra": "^10.0.0",
 		"image-size": "^1.0.0",
 		"png2icons": "^2.0.1",
-		"socket.io": "^4.0.1",
+		"socket.io": "^4.1.1",
 		"uuid": "^8.3.2",
 		"vscode-debugadapter": "^1.47.0",
 		"vscode-debugprotocol": "^1.47.0"
@@ -678,26 +678,26 @@
 		"@types/fs-extra": "^9.0.11",
 		"@types/glob": "^7.1.3",
 		"@types/mocha": "^8.2.2",
-		"@types/node": "^15.0.1",
+		"@types/node": "^15.3.0",
 		"@types/socket.io": "^2.1.13",
 		"@types/uuid": "^8.3.0",
-		"@types/vscode": "^1.55.0",
-		"@typescript-eslint/eslint-plugin": "^4.22.0",
-		"@typescript-eslint/parser": "^4.22.0",
-		"electron": "^12.0.6",
-		"eslint": "^7.25.0",
-		"glob": "^7.1.6",
-		"mocha": "^8.3.2",
-		"ts-loader": "^9.1.1",
+		"@types/vscode": "^1.56.0",
+		"@typescript-eslint/eslint-plugin": "^4.23.0",
+		"@typescript-eslint/parser": "^4.23.0",
+		"electron": "^12.0.7",
+		"eslint": "^7.26.0",
+		"glob": "^7.1.7",
+		"mocha": "^8.4.0",
+		"ts-loader": "^9.1.2",
 		"typescript": "^4.2.4",
 		"vscode-debugadapter-testsupport": "^1.47.0",
 		"vscode-test": "^1.5.2",
-		"webpack": "^5.36.2",
-		"webpack-cli": "^4.6.0"
+		"webpack": "^5.37.0",
+		"webpack-cli": "^4.7.0"
 	},
 	"optionalDependencies": {
 		"bufferutil": "^4.0.3",
-		"utf-8-validate": "^5.0.4"
+		"utf-8-validate": "^5.0.5"
 	},
 	"scripts": {
 		"build": "webpack --config ./core/webpack.config.js",
